+++ index.html
@@ -6,7 +6,7 @@
     <title>NEL VCS</title>
   </head>
 
-  <body>
+  <body style="background-color: black;">
     <div id="app"></div>
     <script type="module" src="/src/main.ts"></script>
   </body>
