+++ src/app.pcss
@@ -12,7 +12,6 @@ body {
 
 body {
   display: grid;
-
   margin: 0;
   font-family: "Inter", sans-serif;
   color: hsl(220 10% 98%);
@@ -27,9 +26,6 @@ body {
   line-height: 1.5;
   font-weight: 400;
   color-scheme: dark;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
-
   font-synthesis: none;
   text-rendering: optimizeLegibility;
   -webkit-font-smoothing: antialiased;
