+++ src/controls/Controls.svelte
@@ -4,14 +4,13 @@
     ConsoleText,
     LocksMap,
     NativeMessage,
-    UI_Controls,
-    UI_StorageFSMs,
-    UI_ClassFSMs
+    UI_Controls
   } from "../stores/stores";
-  import { Smush32 } from "@thi.ng/random";
-  import type { StorageFSM, UI_ControlsMap, UI_Slider } from "../../types";
-  import { get } from "svelte/store";
-  import { updateClassStates } from "../utils/utils";
+
+  import type { UI_Slider } from "../../types";
+
+  import { createEventDispatcher } from "svelte";
+  const dispatch = createEventDispatcher();
 
   function updateControls(e: Event) {
     let { value, dataset } = e.target as HTMLInputElement;
@@ -27,30 +26,6 @@
       );
   }
 
-  function randomisePresets() {
-    const smush = new Smush32(0x909808303);
-    const randomPreset = new Map($UI_Controls);
-
-    $UI_StorageFSMs.forEach((fsm: StorageFSM) => {
-      fsm.randomise();
-      if ( get(fsm) ==='filled')  fsm.storePreset(generateRandomPreset()); 
-    });
- 
-
-    function generateRandomPreset(): UI_ControlsMap {
-      randomPreset.forEach((settings, key) => {
-        let rnd = smush.minmax(0, 1);
-        randomPreset.set(key, {
-          ...settings,
-          value: rnd,
-        });
-      });
-      return randomPreset;
-    }
-    $UI_StorageFSMs = $UI_StorageFSMs;
-    updateClassStates()
-    $UI_ClassFSMs = $UI_ClassFSMs;
-  }
 
 </script>
 
@@ -58,7 +33,7 @@
 
 <div class="sidebar">
   <pre>{$ConsoleText}</pre>
-  <button class="button" on:click={randomisePresets}>Randomise</button>
+  <button class="button" on:click={()=>dispatch('smush')}>Smush</button>
   <h3 class="heading">Controls</h3>
 
   {#if $UI_Controls.size}
